Socket Connection & Server Service boilerplate; Define client-server communication protocols

+++ include/client_networks/matrix.h
@@ -1,7 +1,7 @@
 #ifndef MATRIX_H
 #define MATRIX_H
 #include <stdint.h>
-int32_t determinan(int32_t matrix[3][3]);
-int inverse_matrix(int32_t matrix[3][3], int32_t inversed[3][3]);
-bool is_matrix_equal(int32_t matrix1[3][3], int32_t matrix2[3][3]);
+double determinan(double matrix[3][3]);
+int inverse_matrix(double matrix[3][3], double inversed[3][3]);
+bool is_matrix_equal(double matrix1[3][3], double matrix2[3][3]);
 #endif
